requirements.txt
annotated-types
anyio
blinker
cachetools
certifi
charset-normalizer
click
colorama
Envelopes
Flask
google-auth
openai
h11
httpcore
httpx
idna
itsdangerous
Jinja2
Markdown
markdownmail
MarkupSafe
pillow
pyasn1
pyasn1_modules
pydantic
pydantic_core
requests
rsa
sniffio
supabase
typing-inspection
typing_extensions
urllib3
websockets
Werkzeug